+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.7.0
+
+-   Fixed rvalue row corruption bug on MSVC ([info](https://developercommunity.visualstudio.com/t/C:-Corrupt-references-when-creating-a/10446877))
+-   Added `Cols...` selector template parameter to `at()`, `front()` and `back()`
+
 ## v0.6.0
 
 -   Fixed `swap_columns()` fast-path breaking `data()`